review: net_id needs an "unset", and the C-ABI wiring was untested
Adversarial review returned two blocking defects and three smaller ones.

1. BLOCKING — net_id had no way to say "don't pin one". Config::node.net_id is
   an optional whose nullopt means "unconfigured, accept any net_id", a state
   §3.0 keeps distinct from 0 because collapsing them "made an unconfigured
   node deaf to every non-zero net_id" (rx_node.cpp's own words). Selection
   carried a bare uint8_t, so apply_selection always pinned something and 0 was
   accepted silently. The trap closes through our own API: §15.5a serialises
   the live selection as net_id.value_or(0), so a consumer that reads its
   selection back and feeds it to set_selection pins the filter on 0 and goes
   deaf with no diagnostic. net_id is now int32_t, negative = leave the
   config's value alone, >255 refused.

2. BLOCKING — the C-ABI test could not fail. It pre-stopped the handle before
   running, and the pre-stop return sits ABOVE the load, so config_json and
   the selection were never reached: replacing the whole config-source ternary
   with an unconditional load_all left the suite green (reviewer's mutant,
   reproduced). The fix is a test that reaches the load without a radio — a
   config whose profile_table does not exist fails INSIDE load_finish, after
   the summary is logged and long before any backend — with a wb_log sink
   capturing the summary as evidence. It proves both halves at once: the TEXT
   config was loaded, and the SELECTION was applied to it. The same mutant now
   fails 3 checks.

   That test also required fixing a real defect it exposed: the selection was
   applied AFTER load_finish printed the config summary, so the operator's log
   named the channel the node was configured for rather than the one it would
   fly on. The selection is now threaded into the load and applied first.

3. The TX header's doc block was a verbatim copy of the RX one: it named
   wblink_rx_run and said the refusal returns 2. It returns WBLINK_TX_BAD_ARG
   (-1); 2 in that header's space is the §9.10 wedge a supervisor acts on.

4. Both setters checked arguments before the started-check, so a late call with
   a bad argument returned 2 where the header promises 3. Reordered.

5. node_config_api_test was registered inside the WBLINK_RADIO gate although
   its fixture is air.kind=udp, leaving the contract untested in exactly the
   receive-only builds that motivated a text config. Moved out. The setters'
   own allocation-failure path still used fprintf(stderr) in the pass whose
   point is that stderr is where Android messages go to die; now wb_logf.

ctest 77/77, gates 30/0/0.

+++ node/include/wblink/node/rx_node_c.h
@@ -112,12 +112,26 @@ int wblink_rx_set_config_json(wblink_rx *rx, const char *json);
  * selection that moved only the first would build a diversity node listening
  * in two places for a craft that is only ever in one.
  *
+ * `net_id` IS SIGNED, and negative means "leave the config's value alone".
+ * The config field is an optional whose nullopt means "unconfigured — accept
+ * any net_id", a state §3.0 keeps distinct from 0 because collapsing the two
+ * makes an unconfigured node deaf to every non-zero net_id. 0 is therefore
+ * not a usable stand-in for "unset", and the trap is not hypothetical: §15.5a
+ * serialises the live selection as `net_id.value_or(0)`, so a consumer that
+ * reads a selection back and hands it straight to this call would pin the
+ * filter on 0 and go silently deaf — by a round trip through our own API.
+ * Values above 255 are refused.
+ *
  * `originator` 0 is refused (2) — it is §12's "no preference" sentinel, so a
  * selection naming it is a caller bug rather than a selection — as is a
  * 0 MHz channel, which would tune the node nowhere. Returns 0 on success.
+ *
+ * The selection is applied to the loaded config BEFORE the config summary is
+ * logged, so what an operator reads in the log is the channel the node will
+ * actually fly on.
  */
 int wblink_rx_set_selection(wblink_rx *rx, uint16_t originator,
-                            uint8_t net_id, uint16_t channel_mhz);
+                            int32_t net_id, uint16_t channel_mhz);
 
 /*
  * Supply pre-opened USB device descriptors, parallel to the config's
